inverter aus UI installieren

+++ functions/storm.house.bash
@@ -173,7 +173,9 @@ setup_heatingrod() {
   local cpass
 
 
-   if ! cond_redirect install -m 755 "${sdIncludesDir}/setup_inverter" /usr/local/sbin; then echo "FAILED (install setup_inverter)"; return 1; fi
+  if [[ ! -f /usr/local/sbin/setup_inverter ]]; then
+    if ! cond_redirect install -m 755 "${sdIncludesDir}/setup_inverter" /usr/local/sbin; then echo "FAILED (install setup_inverter)"; return 1; fi
+  fi
 
   for component in things items rules; do
     dir="${OPENHAB_CONF:-/etc/openhab}/${component}/"
